requirements.txt
python-docx==0.8.11
requests==2.31.0
tqdm==4.66.1
python-dotenv==1.0.0
lxml==4.9.3
rich==13.7.0
typer==0.9.0
pydantic==2.5.0
pydantic-settings==2.0.3
tenacity==8.2.3
Pillow==10.1.0
aiohttp==3.9.1 